Simplified Id-referencing

+++ Engine/Core/Components/Game/OwnerIdComponent.cs
@@ -5,6 +5,6 @@ namespace Lockstep.Core.Components.Game
     [Game]
     public class OwnerIdComponent : IComponent
     {                        
-        public int value;
+        public byte value;
     }
 }
